+++ Runtime/Utils/SaintsFieldConfig.cs
@@ -49,19 +49,19 @@ namespace SaintsField.Utils
 
         [Space]
 
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentInChildrenExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentInParentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentInParentsExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentInSceneExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getPrefabWithComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getScriptableObjectExp = EXP.NoPicker | EXP.NoAutoResignToNull;
-        [EnumFlags(defaultExpanded: false)] public EXP getByXPathExp = EXP.None;
+        [FlagsDropdown] public EXP getComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getComponentInChildrenExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getComponentInParentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getComponentInParentsExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getComponentInSceneExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getPrefabWithComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getScriptableObjectExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP getByXPathExp = EXP.None;
 
         [LayoutStart("./Deprecated", ELayout.CollapseBox)]
         [InfoBox("Deprecated", EMessageType.Warning)]
-        [EnumFlags(defaultExpanded: false)] public EXP getComponentByPathExp = EXP.NoAutoResignToValue | EXP.NoAutoResignToNull | EXP.NoPicker;
+        [FlagsDropdown] public EXP getComponentByPathExp = EXP.NoAutoResignToValue | EXP.NoAutoResignToNull | EXP.NoPicker;
         [InfoBox("Deprecated", EMessageType.Warning)]
-        [EnumFlags(defaultExpanded: false)] public EXP findComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
+        [FlagsDropdown] public EXP findComponentExp = EXP.NoPicker | EXP.NoAutoResignToNull;
     }
 }